Moved the metrics collection code out of the listener
